+++ lib/Version.ps1
@@ -1,1 +1,1 @@
-$script:UpdaterVersion = '0.3.0-beta'
+$script:UpdaterVersion = '0.3.1-beta'